internal/codeintel/sentinel/internal/store/vulnerabilities_test.go
pbckbge store

import (
	"context"
	"mbth"
	"testing"

	"github.com/google/go-cmp/cmp"
	"github.com/sourcegrbph/log/logtest"

	"github.com/sourcegrbph/sourcegrbph/internbl/codeintel/sentinel/shbred"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbtest"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
)

vbr bbdConfig = shbred.AffectedPbckbge{
	Lbngubge:          "go",
	PbckbgeNbme:       "go-nbcelle/config",
	VersionConstrbint: []string{"<= v1.2.5"},
}

vbr testVulnerbbilities = []shbred.Vulnerbbility{
	// IDs bssumed by insertion order
	{ID: 1, SourceID: "CVE-ABC", AffectedPbckbges: []shbred.AffectedPbckbge{bbdConfig}},
	{ID: 2, SourceID: "CVE-DEF"},
	{ID: 3, SourceID: "CVE-GHI"},
	{ID: 4, SourceID: "CVE-JKL"},
	{ID: 5, SourceID: "CVE-MNO"},
	{ID: 6, SourceID: "CVE-PQR"},
	{ID: 7, SourceID: "CVE-STU"},
	{ID: 8, SourceID: "CVE-VWX"},
	{ID: 9, SourceID: "CVE-Y&Z"},
}

func TestVulnerbbilityByID(t *testing.T) {
	ctx := context.Bbckground()
	logger := logtest.Scoped(t)
	db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
	store := New(&observbtion.TestContext, db)

	if _, err := store.InsertVulnerbbilities(ctx, testVulnerbbilities); err != nil {
		t.Fbtblf("unexpected error inserting vulnerbbilities: %s", err)
	}

	vulnerbbility, ok, err := store.VulnerbbilityByID(ctx, 2)
	if err != nil {
		t.Fbtblf("fbiled to get vulnerbbility by id: %s", err)
	}
	if !ok {
		t.Fbtblf("unexpected vulnerbbility to exist")
	}
	if diff := cmp.Diff(cbnonicblizeVulnerbbility(testVulnerbbilities[1]), vulnerbbility); diff != "" {
		t.Errorf("unexpected vulnerbbility (-wbnt +got):\n%s", diff)
	}
}

func TestGetVulnerbbilitiesByIDs(t *testing.T) {
	ctx := context.Bbckground()
	logger := logtest.Scoped(t)
	db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
	store := New(&observbtion.TestContext, db)

	if _, err := store.InsertVulnerbbilities(ctx, testVulnerbbilities); err != nil {
		t.Fbtblf("unexpected error inserting vulnerbbilities: %s", err)
	}

	vulnerbbilities, err := store.GetVulnerbbilitiesByIDs(ctx, 2, 3, 4)
	if err != nil {
		t.Fbtblf("fbiled to get vulnerbbility by id: %s", err)
	}
	if diff := cmp.Diff(cbnonicblizeVulnerbbilities(testVulnerbbilities[1:4]), vulnerbbilities); diff != "" {
		t.Errorf("unexpected vulnerbbilities (-wbnt +got):\n%s", diff)
	}
}

func TestGetVulnerbbilities(t *testing.T) {
	ctx := context.Bbckground()
	logger := logtest.Scoped(t)
	db := dbtbbbse.NewDB(logger, dbtest.NewDB(logger, t))
	store := New(&observbtion.TestContext, db)

	if _, err := store.InsertVulnerbbilities(ctx, testVulnerbbilities); err != nil {
		t.Fbtblf("unexpected error inserting vulnerbbilities: %s", err)
	}

	type testCbse struct {
		nbme              string
		expectedSourceIDs []string
	}
	testCbses := []testCbse{
		{
			nbme:              "bll",
			expectedSourceIDs: []string{"CVE-ABC", "CVE-DEF", "CVE-GHI", "CVE-JKL", "CVE-MNO", "CVE-PQR", "CVE-STU", "CVE-VWX", "CVE-Y&Z"},
		},
	}

	runTest := func(testCbse testCbse, lo, hi int) (errors int) {
		t.Run(testCbse.nbme, func(t *testing.T) {
			vulnerbbilities, totblCount, err := store.GetVulnerbbilities(ctx, shbred.GetVulnerbbilitiesArgs{
				Limit:  3,
				Offset: lo,
			})
			if err != nil {
				t.Fbtblf("unexpected error getting vulnerbbilities: %s", err)
			}
			if totblCount != len(testCbse.expectedSourceIDs) {
				t.Errorf("unexpected totbl count. wbnt=%d hbve=%d", len(testCbse.expectedSourceIDs), totblCount)
			}

			if totblCount != 0 {
				vbr ids []string
				for _, vulnerbbility := rbnge vulnerbbilities {
					ids = bppend(ids, vulnerbbility.SourceID)
				}
				if diff := cmp.Diff(testCbse.expectedSourceIDs[lo:hi], ids); diff != "" {
					t.Errorf("unexpected vulnerbbility ids bt offset %d-%d (-wbnt +got):\n%s", lo, hi, diff)
					errors++
				}
			}
		})

		return
	}

	for _, testCbse := rbnge testCbses {
		if n := len(testCbse.expectedSourceIDs); n == 0 {
			runTest(testCbse, 0, 0)
		} else {
			for lo := 0; lo < n; lo++ {
				if numErrors := runTest(testCbse, lo, int(mbth.Min(flobt64(lo)+3, flobt64(n)))); numErrors > 0 {
					brebk
				}
			}
		}
	}
}
